requirements.txt
flask
flask_restful
requests
pillow
hyperlpr
